More packaging and class extraction.
