+++ src/main/java/org/spongepowered/common/event/tracking/IPhaseState.java
@@ -29,14 +29,10 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.WorldServer;
 import org.spongepowered.api.block.BlockSnapshot;
 import org.spongepowered.api.data.Transaction;
-import org.spongepowered.api.entity.Entity;
 import org.spongepowered.api.world.World;
-import org.spongepowered.common.block.SpongeBlockSnapshot;
-import org.spongepowered.common.event.tracking.phase.EntityPhase;
+import org.spongepowered.common.event.tracking.phase.entity.EntityPhase;
 import org.spongepowered.common.event.tracking.phase.TrackingPhase;
 import org.spongepowered.common.world.BlockChange;
-
-import java.util.List;
 
 import javax.annotation.Nullable;
 
@@ -104,5 +100,8 @@ public interface IPhaseState {
 
     default boolean shouldCaptureBlockChangeOrSkip(PhaseContext phaseContext, BlockPos pos) {
         return true;
+    }
+    default boolean isInteraction() {
+        return false;
     }
 }
