+++ src/main/java/org/spongepowered/common/event/tracking/phase/PlayerPhase.java
@@ -44,8 +44,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import javax.annotation.Nullable;
-
 public class PlayerPhase extends TrackingPhase {
 
     public static final class State {
