+++ src/main/java/org/spongepowered/common/event/tracking/phase/entity/PlayerWakeUpState.java
@@ -22,26 +22,20 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.event.tracking.phase;
+package org.spongepowered.common.event.tracking.phase.entity;
 
-import org.spongepowered.common.event.tracking.phase.entity.EntityPhase;
-import org.spongepowered.common.event.tracking.phase.general.GeneralPhase;
-import org.spongepowered.common.event.tracking.phase.generation.GenerationPhase;
-import org.spongepowered.common.event.tracking.phase.packet.PacketPhase;
-import org.spongepowered.common.event.tracking.phase.tick.TickPhase;
+import org.spongepowered.common.event.tracking.CauseTracker;
+import org.spongepowered.common.event.tracking.PhaseContext;
+import org.spongepowered.common.event.tracking.TrackingUtil;
 
-public final class TrackingPhases {
+class PlayerWakeUpState extends EntityPhaseState {
 
-    public static final TickPhase TICK = TickPhase.getInstance();
-    public static final EntityPhase ENTITY = EntityPhase.getInstance();
-    public static final BlockPhase BLOCK    = BlockPhase.getInstance();
-    public static final GeneralPhase GENERAL  = GeneralPhase.getInstance();
-    public static final PacketPhase PACKET = PacketPhase.getInstance();
-    public static final PluginPhase PLUGIN = PluginPhase.getInstance();
-    public static final GenerationPhase GENERATION = GenerationPhase.getInstance();
-    public static final PlayerPhase PLAYER = PlayerPhase.getInstance();
-
-    private TrackingPhases() {
+    PlayerWakeUpState() {
     }
 
+    @Override
+    void unwind(CauseTracker causeTracker, PhaseContext context) {
+        context.getCapturedBlockSupplier().ifPresentAndNotEmpty(blocks -> TrackingUtil
+                .processBlockCaptures(blocks, causeTracker, this, context));
+    }
 }
